% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_price.R
\name{load_price}
\alias{load_price}
\title{load price data for live hirame in Tokyo Central Market}
\usage{
load_price()
}
\value{
tibble with price of live hirame in Tokyo
}
\description{
load price data for live hirame in Tokyo Central Market
}
\examples{
\dontrun{
load_price()
}
}
