% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prefec_eng}
\alias{prefec_eng}
\title{Translate Prefec name to English}
\usage{
prefec_eng(dat)
}
\arguments{
\item{dat}{Location vector}
}
\value{
English-translated prefec name
}
\description{
Translate Prefec name to English
}
