% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_hiramelive.R
\name{read_hiramelive}
\alias{read_hiramelive}
\title{Read hirame (live) market stats for Tokyo central market}
\usage{
read_hiramelive(fname)
}
\arguments{
\item{fname}{file name}
}
\value{
tibble with hirame quantity, total_price, and average_price
}
\description{
Read hirame (live) market stats for Tokyo central market
}
\examples{
\dontrun{
read_hiramelive("../data-raw/excel/2002_hiramelive.csv")
}
}
