% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sort_prefec}
\alias{sort_prefec}
\title{Sort prefecture in southward order}
\usage{
sort_prefec(dat)
}
\arguments{
\item{dat}{Prefec vector}
}
\value{
factor with prefec levels
}
\description{
Sort prefecture in southward order
}
